+++ app/core/erp/class_views/categoryMaterials/views.py
@@ -12,6 +12,8 @@ from django.views.generic import ListView, CreateView, UpdateView, DeleteView
 from django.urls import reverse_lazy
 from django.http import JsonResponse
 from django.views.generic import TemplateView
+from django.contrib.auth.models import Group
+
 
 
 
@@ -38,16 +40,22 @@ class categoryMaterialsListView(LoginRequiredMixin, TemplateView):
                 cat.user_create = request.user.username
                 cat.save()
             elif action == 'edit':
-                cat = CategoryMaterials.objects.get(pk=request.POST['id'])
-                cat.name = request.POST['name']
-                cat.unity = request.POST['unity']
-                cat.user_update = request.user.username
-                cat.save()
+                if request.session['group'] == Group.objects.get(pk=1):
+                    cat = CategoryMaterials.objects.get(pk=request.POST['id'])
+                    cat.name = request.POST['name']
+                    cat.unity = request.POST['unity']
+                    cat.user_update = request.user.username
+                    cat.save()
+                else:
+                    data['error'] = 'No tiene permiso para ingresar a este módulo'
             elif action == 'delete':
-                cat = CategoryMaterials.objects.get(pk=request.POST['id'])
-                cat.user_update = request.user.username
-                cat.save()
-                cat.delete()
+                if request.session['group'] == Group.objects.get(pk=1):
+                    cat = CategoryMaterials.objects.get(pk=request.POST['id'])
+                    cat.user_update = request.user.username
+                    cat.save()
+                    cat.delete()
+                else:
+                    data['error'] = 'No tiene permiso para ingresar a este módulo'
 
             else:
                 data['error'] = 'Ha ocurrido un error'
