+++ app/core/erp/class_views/production/views.py
@@ -8,6 +8,8 @@ from django.urls import reverse_lazy
 from django.utils.decorators import method_decorator
 from django.views import View
 from django.views.decorators.csrf import csrf_exempt
+from django.contrib.auth.models import Group
+
 
 from core.erp.forms import BuyForm, ProductForm, ProductionForm, SupplierForm
 from django.views.generic import CreateView, ListView, DeleteView, UpdateView
@@ -44,13 +46,16 @@ class ProductionListView(LoginRequiredMixin, ListView):
                 for i in DetProduction.objects.filter(crea_id=request.POST['id']):  
                     data.append(i.toJSON())  
             elif action == 'delete':
-                cli = Production.objects.get(pk=request.POST['id'])
-                cli.user_update = request.user.username
-                pro = Product.objects.get(id = cli.produc_id)
-                pro.stock -= (cli.total)
-                pro.save()
-                cli.save()
-                cli.delete()
+                if request.session['group'] == Group.objects.get(pk=1):
+                    cli = Production.objects.get(pk=request.POST['id'])
+                    cli.user_update = request.user.username
+                    pro = Product.objects.get(id = cli.produc_id)
+                    pro.stock -= (cli.total)
+                    pro.save()
+                    cli.save()
+                    cli.delete()
+                else:
+                    data['error'] = 'No tiene permiso para ingresar a este módulo'
 
             else:
                 data['error'] = 'Ha ocurrido un error'
